+++ changelog.markdown
@@ -1,5 +1,10 @@
 Mar 24, 2013
 ============
+foxiepaws - more core changes, addition of module namespace
+  * Modules: First revision of the ShadowIRCd protocol module
+  * Core: added in snotice handling using the protocol module
+  * Core: added in log channel area (not yet logging to channel though!)
+  * Config: added in channel list and removed the rest of my personal config data
 foxiepaws - Heavy changes to the core, Pretty Much an entire rewrite
 
   * Now using GetOpt to allow passing config files and verbose/debug modes